organized timestepping files
