+++ dymos/transcriptions/timestepping/euler_integration_comp.py
@@ -2,253 +2,8 @@ import collections
 
 import numpy as np
 import openmdao.api as om
-import dymos as dm
 
-class SimpleODE(om.ExplicitComponent):
-    """
-    A simple ODE from https://math.okstate.edu/people/yqwang/teaching/math4513_fall11/Notes/rungekutta.pdf
-    """
-    def initialize(self):
-        self.options.declare('num_nodes', types=(int,))
-
-    def setup(self):
-        nn = self.options['num_nodes']
-        self.add_input('x', shape=(nn,), units='s**2')
-        self.add_input('t', shape=(nn,), units='s')
-
-        self.add_output('x_dot', shape=(nn,), units='s')
-
-        ar = np.arange(nn, dtype=int)
-        self.declare_partials(of='x_dot', wrt='x', rows=ar, cols=ar, val=1.0)
-        self.declare_partials(of='x_dot', wrt='t', rows=ar, cols=ar)
-
-    def compute(self, inputs, outputs):
-        x = inputs['x']
-        t = inputs['t']
-        outputs['x_dot'] = x - t**2 + 1
-
-    def compute_partials(self, inputs, partials):
-        t = inputs['t']
-        partials['x_dot', 't'] = -2*t
-
-
-class ODEEvaluationGroup(om.Group):
-    """
-    A special group whose purpose is to evaluate the ODE and return the computed
-    state weights.
-    """
-
-    def __init__(self, ode_class, time_options, state_options, control_options,
-                 polynomial_control_options, parameter_options, ode_init_kwargs=None):
-        super().__init__()
-
-        # Get the state vector.  This isn't necessarily ordered
-        # so just pick the default ordering and go with it.
-        self.state_options = state_options
-        self.time_options = time_options
-        self.control_options = control_options
-        self.polynomial_control_options = polynomial_control_options
-        self.parameter_options = parameter_options
-        self.control_interpolants = {}
-        self.polynomial_control_interpolants = {}
-        self.ode_class = ode_class
-        self.ode_init_kwargs = {} if ode_init_kwargs is None else ode_init_kwargs
-
-    def setup(self):
-        if self.control_options:
-            # Add control interpolant
-            raise NotImplementedError('dynamic controls not yet implemented')
-        if self.polynomial_control_options:
-            # Add polynomial control interpolant
-            raise NotImplementedError('polynomial controls not yet implemented')
-
-        self.add_subsystem('ode', self.ode_class(num_nodes=1, **self.ode_init_kwargs))
-
-        self.add_subsystem('state_rate_collector',
-                           StateRateCollectorComp(state_options=self.state_options,
-                                                  time_units=self.time_options['units']))
-
-    def configure(self):
-        self._configure_time()
-        self._configure_states()
-
-    def _configure_time(self):
-        targets = self.time_options['targets']
-        time_phase_targets = self.time_options['time_phase_targets']
-        t_initial_targets = self.time_options['t_initial_targets']
-        t_duration_targets = self.time_options['t_duration_targets']
-        units = self.time_options['units']
-
-        for tgts, var in [(targets, 'time'), (time_phase_targets, 'time_phase'),
-                          (t_initial_targets, 't_initial'), (t_duration_targets, 't_duration')]:
-            for t in tgts:
-                self.promotes('ode', inputs=[(t, var)])
-                print(f'promoted {t} to {var}')
-            if tgts:
-                self.set_input_defaults(name=var,
-                                        val=np.ones((1,)),
-                                        units=units)
-
-    def _configure_states(self):
-        for name, options in self.state_options.items():
-            shape = options['shape']
-            targets = options['targets']
-            rate_src = options['rate_source']
-            rate_path, rate_io = self._get_rate_source_path(name)
-
-            # Promote targets from the ODE
-            for tgt in targets:
-                self.promotes('ode', inputs=[(tgt, f'states:{name}')])
-            if targets:
-                self.set_input_defaults(name=f'states:{name}',
-                                        val=np.ones(shape),
-                                        units=options['units'])
-
-            # If the state rate source is an output, connect it, otherwise
-            # promote it to the appropriate name
-            if rate_io == 'output':
-                self.connect(rate_path, f'state_rate_collector.state_rates_in:{name}_rate')
-            else:
-                self.promotes('state_rate_collector',
-                              inputs=[(f'state_rates_in:{name}_rate', rate_path)])
-
-            self.add_design_var(f'states:{name}')
-            self.add_constraint(f'state_rate_collector.state_rates:{name}_rate')
-
-    def _get_rate_source_path(self, state_var):
-        """
-        Get path of the rate source variable so that we can connect it to the
-        outputs when we're done.
-
-        Parameters
-        ----------
-        state_var : str
-            The name of the state variable whose path is desired.
-
-        Returns
-        -------
-        path : str
-            The path to the rate source of the state variable.
-        io : str
-            A string indicating whether the variable in the path is an 'input'
-            or an 'output'.
-        """
-        var = self.state_options[state_var]['rate_source']
-
-        if var == 'time':
-            rate_path = 'time'
-            io = 'input'
-        elif var == 'time_phase':
-            rate_path = 'time_phase'
-            io = 'input'
-        elif self.state_options is not None and var in self.state_options:
-            rate_path = f'states:{var}'
-            io = 'input'
-        elif self.control_options is not None and var in self.control_options:
-            rate_path = f'controls:{var}'
-            io = 'output'
-        elif self.polynomial_control_options is not None and var in self.polynomial_control_options:
-            rate_path = f'polynomial_controls:{var}'
-            io = 'output'
-        elif self.parameter_options is not None and var in self.parameter_options:
-            rate_path = f'parameters:{var}'
-            io = 'input'
-        elif var.endswith('_rate') and self.control_options is not None and \
-                var[:-5] in self.control_options:
-            rate_path = f'control_rates:{var}'
-            io = 'output'
-        elif var.endswith('_rate2') and self.control_options is not None and \
-                var[:-6] in self.control_options:
-            rate_path = f'control_rates:{var}'
-            io = 'output'
-        elif var.endswith('_rate') and self.polynomial_control_options is not None and \
-                var[:-5] in self.polynomial_control_options:
-            rate_path = f'polynomial_control_rates:{var}'
-            io = 'output'
-        elif var.endswith('_rate2') and self.polynomial_control_options is not None and \
-                var[:-6] in self.polynomial_control_options:
-            rate_path = f'polynomial_control_rates:{var}'
-            io = 'output'
-        else:
-            rate_path = f'ode.{var}'
-            io = 'output'
-        return rate_path, io
-
-
-from dymos.utils.misc import get_rate_units
-from dymos.options import options as dymos_options
-
-
-class StateRateCollectorComp(om.ExplicitComponent):
-    """
-    Class definition for StateRateCollectorComp.
-
-    Collects the state rates and outputs them in the units specified in the state options.
-    For explicit integration this is necessary when the output providing the state rate has
-    different units than those defined in the state_options/time_options.
-
-    Parameters
-    ----------
-    **kwargs : dict
-        Dictionary of optional arguments.
-    """
-
-    def initialize(self):
-        """
-        Declare component options.
-        """
-        self.options.declare(
-            'state_options', types=dict,
-            desc='Dictionary of options for the ODE state variables.')
-        self.options.declare(
-            'time_units', default=None, allow_none=True, types=str,
-            desc='Units of time')
-
-        # Save the names of the dynamic controls/parameters
-        self._input_names = {}
-        self._output_names = {}
-
-        self._no_check_partials = not dymos_options['include_check_partials']
-
-    def setup(self):
-        """
-        Create inputs/outputs on this component.
-        """
-        state_options = self.options['state_options']
-        time_units = self.options['time_units']
-
-        for name, options in state_options.items():
-            self._input_names[name] = f'state_rates_in:{name}_rate'
-            self._output_names[name] = f'state_rates:{name}_rate'
-            shape = options['shape']
-            size = np.prod(shape, dtype=int)
-            units = options['units']
-
-            rate_units = get_rate_units(units, time_units)
-
-            self.add_input(self._input_names[name], shape=shape, units=rate_units)
-            self.add_output(self._output_names[name], shape=shape, units=rate_units)
-
-            ar = np.arange(size, dtype=int)
-            self.declare_partials(of=self._output_names[name],
-                                  wrt=self._input_names[name],
-                                  rows=ar, cols=ar, val=1.0)
-
-    def compute(self, inputs, outputs):
-        """
-        Compute component outputs.
-
-        Parameters
-        ----------
-        inputs : `Vector`
-            `Vector` containing inputs.
-        outputs : `Vector`
-            `Vector` containing outputs.
-        """
-        state_options = self.options['state_options']
-
-        for name, options in state_options.items():
-            outputs[self._output_names[name]] = inputs[self._input_names[name]]
+from .ode_evaluation_group import ODEEvaluationGroup
 
 
 class EulerIntegrationComp(om.ExplicitComponent):
@@ -749,123 +504,3 @@ class EulerIntegrationComp(om.ExplicitComponent):
 
             self._propagate(inputs, outputs=False, derivs=None, time_stack=time_stack, state_stack=state_stack)
             self._backpropagate_derivs(inputs, partials, time_stack=time_stack, state_stack=state_stack)
-
-def test_eval():
-    ode_class = SimpleODE
-    time_options = dm.phase.options.TimeOptionsDictionary()
-
-    time_options['targets'] = 't'
-    time_options['units'] = 's'
-
-    state_options = {'x': dm.phase.options.StateOptionsDictionary()}
-
-    state_options['x']['shape'] = (1,)
-    state_options['x']['units'] = 's**2'
-    state_options['x']['rate_source'] = 'x_dot'
-    state_options['x']['targets'] = ['x']
-
-    control_options = {}
-    polynomial_control_options = {}
-    parameter_options = {}
-
-    p = om.Problem()
-    p.model.add_subsystem('ode_eval', ODEEvaluationGroup(ode_class, time_options, state_options,
-                                                         control_options,
-                                                         polynomial_control_options,
-                                                         parameter_options, ode_init_kwargs=None))
-    p.setup()
-
-    p.set_val('ode_eval.states:x', [1.25])
-    p.set_val('ode_eval.time', [2.2])
-
-    p.run_model()
-
-    p.model.list_inputs()
-    p.model.list_outputs()
-    # p.check_partials()
-
-
-
-def test_fwd():
-    ode_class = SimpleODE
-    time_options = dm.phase.options.TimeOptionsDictionary()
-
-    time_options['targets'] = 't'
-    time_options['units'] = 's'
-
-    state_options = {'x': dm.phase.options.StateOptionsDictionary()}
-
-    state_options['x']['shape'] = (1,)
-    state_options['x']['units'] = 's**2'
-    state_options['x']['rate_source'] = 'x_dot'
-    state_options['x']['targets'] = ['x']
-
-    control_options = {}
-    polynomial_control_options = {}
-    parameter_options = {}
-
-    p = om.Problem()
-
-    p.model.add_subsystem('fixed_step_integrator', EulerIntegrationComp(ode_class, time_options, state_options,
-                                                                        parameter_options, control_options,
-                                                                        polynomial_control_options, mode='fwd',
-                                                                        num_steps=100,
-                                                                        ode_init_kwargs=None))
-    p.setup(mode='fwd', force_alloc_complex=True)
-
-    p.set_val('fixed_step_integrator.state_initial_value:x', 0.5)
-    p.set_val('fixed_step_integrator.t_initial', 0.0)
-    p.set_val('fixed_step_integrator.t_duration', 2.0)
-
-    p.run_model()
-
-    print('t_f', p.get_val('fixed_step_integrator.t_final'))
-    print('x_f', p.get_val('fixed_step_integrator.state_final_value:x'))
-
-    p.check_partials(method='fd', form='central')
-
-
-def test_rev():
-    ode_class = SimpleODE
-    time_options = dm.phase.options.TimeOptionsDictionary()
-
-    time_options['targets'] = 't'
-    time_options['units'] = 's'
-
-    state_options = {'x': dm.phase.options.StateOptionsDictionary()}
-
-    state_options['x']['shape'] = (1,)
-    state_options['x']['units'] = 's**2'
-    state_options['x']['rate_source'] = 'x_dot'
-    state_options['x']['targets'] = ['x']
-
-    control_options = {}
-    polynomial_control_options = {}
-    parameter_options = {}
-
-    p = om.Problem()
-
-    p.model.add_subsystem('fixed_step_integrator', EulerIntegrationComp(ode_class, time_options, state_options,
-                                                                        parameter_options, control_options,
-                                                                        polynomial_control_options, mode='rev',
-                                                                        num_steps=10,
-                                                                        ode_init_kwargs=None))
-    p.setup(mode='rev', force_alloc_complex=True)
-
-    p.set_val('fixed_step_integrator.state_initial_value:x', 0.5)
-    p.set_val('fixed_step_integrator.t_initial', 0.0)
-    p.set_val('fixed_step_integrator.t_duration', 2.0)
-
-    p.run_model()
-
-    print('t_f', p.get_val('fixed_step_integrator.t_final'))
-    print('x_f', p.get_val('fixed_step_integrator.state_final_value:x'))
-
-    p.check_partials(method='fd', form='central')
-
-
-
-if __name__ == '__main__':
-    # test_eval()
-    test_fwd()
-    # test_rev()
